# Small clean-ups for the random module