Timings:

```
$ python3.9 -m timeit -r 11 -s 'from math import floor' -s '_int = int' -s 'x=3.14' '_int(x)'
5000000 loops, best of 11: 68 nsec per loop
$ python3.9 -m timeit -r 11 -s 'from math import floor' -s '_int = int' -s 'x=3.14' 'floor(x)'
10000000 loops, best of 11: 36.7 nsec per loop
```