@rhettinger 2020-06-22
The default argument ``_int = int`` is no longer of much value.  The global variable lookup is now slightly cheaper than the cost of the extra default argument for the first lookup.

Baseline timings:

```
$ pytime -s 'from random import randrange' 'randrange(1000)'
500000 loops, best of 11: 559 nsec per loop
$ pytime -s 'from random import randrange' 'randrange(1000, 2000)'
500000 loops, best of 11: 744 nsec per loop
$ pytime -s 'from random import randrange' 'randrange(1000, 2000, 100)'
200000 loops, best of 11: 1.04 usec per loop
```

New timings:
```
$ pytime -s 'from random import randrange' 'randrange(1000)'
500000 loops, best of 11: 556 nsec per loop
$ pytime -s 'from random import randrange' 'randrange(1000, 2000)'
500000 loops, best of 11: 747 nsec per loop
$ pytime -s 'from random import randrange' 'randrange(1000, 2000, 100)'
200000 loops, best of 11: 1.05 usec per loop
```